A little bit of work for future enhanced packet system
