+++ src/ClassicUO.Client/Configuration/Settings.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using ClassicUO.Configuration.Json;
+using ClassicUO.Game;
 using Microsoft.Xna.Framework;
 
 namespace ClassicUO.Configuration
@@ -87,19 +88,6 @@ namespace ClassicUO.Configuration
 
         [JsonPropertyName("plugins")] public string[] Plugins { get; set; } = { "" };
 
-        public bool EnhancedPacketsEnabled = PacketsEnabled();
-
-        private static bool PacketsEnabled()
-        {
-            //Disable enhanced packets if the file exists
-            //Can't put it in user profile folder because we need it's value before we load profiles
-            //Can't put in this global settings JSON because it may mess up launchers
-            if (File.Exists(Path.Combine(CUOEnviroment.ExecutablePath, "Data", "DISABLE_ENHANCED_PACKETS")))
-                return false;
-
-            return true;
-        }
-        
         public static string GetSettingsFilepath()
         {
             if (CustomSettingsFilepath != null)
