+++ src/ClassicUO.Client/Network/EnhancedPackets/EnhancedPacketTypeEnum.cs
